@@ -1,27 +1,27 @@
 {
-    "name": "jenssegers/agent",
-    "description": "Desktop/mobile user agent parser with support for Laravel, based on Mobiledetect",
+    "name": "sureshinde/mobile-desktop-detect",
+    "description": "Desktop/mobile user agent parser with support for Laravel, based on MobiledDetect",
     "keywords": ["laravel", "useragent", "agent", "user agent", "browser", "platform", "mobile", "desktop"],
-    "homepage": "https://github.com/jenssegers/agent",
+    "homepage": "https://github.com/sureshinde/mobile-desktop-detect",
     "license" : "MIT",
     "authors": [
         {
-            "name": "Jens Segers",
-            "homepage": "https://jenssegers.com"
+            "name": "Suresh Shinde",
+            "homepage": "https://sureshinde.com"
         }
     ],
     "require": {
-        "php": ">=5.6",
-        "mobiledetect/mobiledetectlib": "^2.7.6",
+        "php": ">=8.0",
+        "mobiledetect/mobiledetectlib": "^4.8",
         "jaybizzle/crawler-detect": "^1.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.0|^6.0|^7.0",
+        "phpunit/phpunit": "^9.6",
         "php-coveralls/php-coveralls": "^2.1"
     },
     "autoload": {
         "psr-4": {
-            "Jenssegers\\Agent\\": "src/"
+            "SureShinde\\MobileDesktopDetect\\Agent\\": "src/"
         }
     },
     "extra": {
@@ -30,10 +30,10 @@
         },
         "laravel": {
             "providers": [
-                "Jenssegers\\Agent\\AgentServiceProvider"
+                "SureShinde\\MobileDesktopDetect\\AgentServiceProvider"
             ],
             "aliases": {
-                "Agent": "Jenssegers\\Agent\\Facades\\Agent"
+                "Agent": "SureShinde\\MobileDesktopDetect\\Facades\\Agent"
             }
         }
     },
